RMP and comments
