+++ src/pages/detail/[id].js
@@ -1,32 +1,46 @@
-import { Inter } from 'next/font/google'
-import { semesters, subjects } from "../../lib/utils"
-const inter = Inter({ subsets: ['latin'] })
-import { useRouter } from 'next/router';
-import { useEffect, useState, useRef } from 'react';
-import ErrorPage from 'next/error'
+/*
+ * Details page for a specific course - includes course description, instructors, GPA, RMP ratings,
+ * prerequisites, and a calendar view of lecture times.
+ */
 
+// Next.js imports -----
+import { Inter } from 'next/font/google'
+const inter = Inter({ subsets: ['latin'] })
+
+import Head from 'next/head';
+import Script from 'next/script';
+import { useRouter } from 'next/router';
+import ErrorPage from 'next/error'
+// ---------------------
+
+// React imports
+import { useEffect, useState, useRef } from 'react';
+
+// Misc imports -----
 import Select from 'react-select';
 
-import { Image, Icon } from '@chakra-ui/react'
+import { Image, Icon, Spinner } from '@chakra-ui/react'
 import { FaHome, FaInfo } from "react-icons/fa";
 
 import {
   CircularProgressbar, buildStyles
 } from "react-circular-progressbar";
 import "react-circular-progressbar/dist/styles.css";
+// ------------------
 
-
-
+// Component imports -----
 import { instructorStyles, graphColors, boilerExamsCourses, labels } from '@/lib/utils';
+import { semesters, subjects } from "@/lib/utils"
+
 import Footer from '@/components/footer';
-import Head from 'next/head';
-import Calendar from '../../components/calendar';
-import Graph from '../../components/graph';
+import Calendar from '@/components/calendar';
+import Graph from '@/components/graph';
 import GpaModal from '@/components/gpaModal';
 import InfoModal from '@/components/infoModal';
 import FullInstructorModal from '@/components/fullInstructorModal';
 import Prereqs from '@/components/prereqs';
-import Script from 'next/script';
+import { loadRatingsForProfs } from '@/components/RMP';
+// ------------------------
 
 
 const CardDetails = ({ courseData, semData }) => {
@@ -110,13 +124,6 @@ const CardDetails = ({ courseData, semData }) => {
 
 
     // Set selectable instructors
-    // const instrs = [];
-    // for (const instructor in course.gpa) {
-    //   if (!instrs.includes(instructor)) {
-    //     instrs.push(instructor);
-    //   }
-    // }
-    // console.log("instructors: " + instrs)
     setSelectableInstructors(allProfs);
 
     // set current semester
@@ -146,44 +153,16 @@ const CardDetails = ({ courseData, semData }) => {
 
   // Another UseEffect to asynchronously get RMP ratings
   useEffect(() => {
-    if (!course) return;
-
-    const allProfs = [];
-    for (const semester in course.instructor) {
-      for (const instructor of course.instructor[semester]) {
-        if (!allProfs.includes(instructor)) {
-          allProfs.push(instructor);
-        }
-      }
-    }
-
-    async function loadRatingsForProfessors(allProfs) {
-      try {
-        if (isLoadingRatings.current) {
-          return;
-        }
-
-        isLoadingRatings.current = true;
-
-        const ratings = await getAllRMPRatings(allProfs);
-        setCurRMP((prevRMP) => {
-          return { ...prevRMP, ...ratings };
-        });
-      } catch (error) {
-        console.error("Error loading ratings:", error);
-      } finally {
-        isLoadingRatings.current = false;
-      }
-    }
-
-    loadRatingsForProfessors(allProfs);
+    loadRatingsForProfs(course).then((ratings) => {
+      setCurRMP(ratings);
+    });
   }, [course]);
 
 
+  // ------------------ STATES ------------------ //
   const [firstInstructor, setFirstInstructor] = useState("");
   const [curGPA, setCurGPA] = useState({});
   const [curRMP, setCurRMP] = useState({});
-  const isLoadingRatings = useRef(false);
 
   const [sem, setSem] = useState(semData);
   const [gpaGraph, setGpaGraph] = useState({});
@@ -208,24 +187,8 @@ const CardDetails = ({ courseData, semData }) => {
     } catch { }
   });
 
-  // Function for changing instructors on semester select (old feature)
-  // const changeInstructors = (sem) => {
 
-  //   setSem(sem);
-  //   availableSemesters.forEach((otherSem) => {
-  //     if (otherSem !== sem) {
-  //       try {
-  //         document.getElementById(otherSem).classList.remove("bg-sky-600");
-  //       } catch { }
-  //     }
-  //   });
-
-  //   try {
-  //     document.getElementById(sem).classList.add("bg-sky-600");
-  //   } catch { }
-  // }
-
-
+  // Function to get searchable prof string for Reddit search
   const getSearchableProfString = () => {
     //create ret string = ""
     //for each prof in curInstructors, add to ret string with " OR "
@@ -296,89 +259,18 @@ const CardDetails = ({ courseData, semData }) => {
   }
 
 
-  // Function to get link for RMP on the dial graph
-  // function formatInstructorNameRMP(instructor) {
-  //   if (!instructor) return ''; // Check if instructor is provided
-  //   const nameParts = instructor.split(", ");
-  //   if (nameParts.length < 2) return ''; // Check if split operation succeeded
-  //   const firstName = nameParts[1].split(" ")[0];
-  //   const splitName = nameParts[0].split(" ");
-  //   const lastName = splitName[splitName.length - 1];
-
-  //   // splits from "Middle middle Last, First" to "First Last"
-  //   return `${firstName} ${lastName}`;
-  // }
-
-
-  // Batched RMP ratings fetch for allProfs
-  async function getAllRMPRatings(allProfs) {
-    if (!Array.isArray(allProfs) || allProfs.length === 0) return {};
-
-    const batchSize = 10; // Number of requests to send in each batch
-    const batches = []; // Array to hold batches of professors
-    const ratings = {}; // Object to store ratings
-
-    // Split allProfs into batches
-    for (let i = 0; i < allProfs.length; i += batchSize) {
-      batches.push(allProfs.slice(i, i + batchSize));
-    }
-
-    // Process batches in parallel
-    await Promise.all(
-      batches.map(async (batch) => {
-        const ratingsBatch = await Promise.all(batch.map((instructor) => getRMPRating(instructor)));
-        batch.forEach((instructor, index) => {
-          ratings[instructor] = ratingsBatch[index];
-        });
-      })
-    );
-
-    return ratings;
-  }
-
-
-  // Get RateMyProfessor ratings for instructor
-  async function getRMPRating(instructor) {
-    if (!instructor) return 0;
-
-    try {
-      // TO SEARCH FOR "PURDUE UNIVERSITY"
-      // const params = new URLSearchParams({ q: "Purdue University" });
-      // const responseSchools = await fetch("/api/ratings/searchSchool?" + params);
-      // const schools = await responseSchools.json();
-      // const purdues = schools.schools.filter(school => school.city === "West Lafayette");
-
-      let rating = 0;
-
-      // for all Purdue University schools in West Lafayette, search prof
-      const schools = ["U2Nob29sLTc4Mw==", "U2Nob29sLTE3NTk5"]; // purdue IDs for West Lafayette
-      for (const school of schools) {
-        const paramsTeacher = new URLSearchParams({ name: instructor, id: school });
-        const responseProf = await fetch("/api/ratings/searchTeacher?" + paramsTeacher);
-        const prof = await responseProf.json();
-        const profs = prof.prof.filter(Boolean);
-
-        if (profs.length > 0) {
-          const paramsGetTeacher = new URLSearchParams({ id: profs[0].id });
-          const responseRMP = await fetch("/api/ratings/getTeacher?" + paramsGetTeacher);
-          const RMPrating = await responseRMP.json();
-          rating = RMPrating.RMPrating.avgRating;
-          break;
-        }
-      }
-
-      return rating;
-    } catch (error) {
-      console.error(error);
-      return 0;
-    }
-  }
-
-
   ///////////////////////////////////////  RENDER  /////////////////////////////////////////
 
   if (JSON.stringify(course) == '{}') {
     return <ErrorPage statusCode={404} />
+  }
+
+  if (loading) {
+    return (
+      <div className='h-screen w-screen flex items-center justify-center'>
+        <Spinner size="lg" color="white" />
+      </div>
+    )
   }
 
   return (
@@ -514,18 +406,6 @@ const CardDetails = ({ courseData, semData }) => {
               </div>
             </div>
 
-
-            {/* Semester Tags */}
-            {/* <div className="flex flex-row flex-wrap gap-1 mb-1">
-              {availableSemesters.map((sem, i) => (
-                <button className={`text-xs px-2 py-1 rounded-full border-solid border
-                                          ${i === 0 ? "bg-sky-600" : ""} border-sky-800 whitespace-nowrap transition-all`}
-                  key={i}
-                  id={sem}
-                  onClick={() => changeInstructors(sem)}>{sem}</button>
-              ))}
-            </div>
-             */}
 
             {/* Other Links Buttons */}
             <div className="flex flex-row flex-wrap my-2">
@@ -682,6 +562,8 @@ const CardDetails = ({ courseData, semData }) => {
 };
 
 
+
+// @Sarthak made this, some api call to get course data
 export async function getServerSideProps(context) {
 
   const params = new URLSearchParams({ detailId: context.params.id });
